[v0.9.8.1-qfit3-resol1-cln3-newf24] Fix size_t bug, use uint64_t to multiply in 32 bits

+++ pgnget.c
@@ -372,7 +372,7 @@ pgn_result_collect (struct pgn_result *p)
 }
 #endif
 
-	ok = ok && (size_t)DaBa.n_games < ((size_t)MAXGAMESxBLOCK*(size_t)MAXBLOCKS);
+	ok = ok && (uint64_t)DaBa.n_games < ((uint64_t)MAXGAMESxBLOCK*(uint64_t)MAXBLOCKS);
 
 	if (ok) {
